Fix order-dependence for @only_once by doing Handler registration
The order in which the @only_once decorator is applied was significant
because it had to be applied before Handler registration to be
effective, meaning it **had** to be the inner-most decorator.

Now, it uses Handler registration just like @when, etc.  This means the
order doesn't matter, but it does mean that it can't be used on normal,
non-reactive-handler functions.