Fix cold loading of relation instances
From Alex's debugging investigation, it seems that the import loading
refactor (#51) made the assumption that handler discovery will always
have been run before RelationBase.from_state() or
RelationBase.from_name() are called.  For actions or the CLI, this won't
be true, and we may need to attempt to import the appropriate file.

Fixes #105