Rename Endpoint.joined to Endpoint.is_joined
It came up that an interface layer could easily want to name a handler
`joined` which ends up conflicting with the property and breaking the
management of the `endpoint.{endpoint_name}.joined` flag causing it to
always be set, even when no related units are joined.

The `is_joined` name is more descriptive to indicate that it's a boolean
flag, anyway.