Manage automatic flags in the library
This pulls in the automatic flags that are currently managed by the
leadership and basic layer, as well as adding support for flags tracking
the upgrade of charm, resource, and OS series.  It also adds a docs page
documenting all of the automatically managed flags in one place, as well
as the recommended conventions for interacting with them.