Add ability to trigger on flag being cleared
Triggers currently allow another flag to be set or cleared when a given
flag is set.  This happens as soon as the flag is set, before any
handlers are invoked, which is important for ensuring that related flags
are handled properly.

This adds the ability to trigger another flag being set or cleared when
a given flag is cleared, as well, which addresses race conditions such
as the one reported in juju-solutions/interface-public-address#4.