Fix errant str.format handling of flags in expand_name (#210)
* Fix errant str.format handling of flags in expand_name

If a flag contains format string constructs, especially `{}`, then
`Endpoint.expand_name` will blow up with a very misleading error. Doing
this is probably a charm bug, but it is easier to debug if you can see
the flag with the unprocessed format string constructs rather than a
random seeming `IndexError`. This changes it to only replace the exact
string `{endpoint_name}` to prevent the error.

* Fix lint error from unused import