Add comment about flags executing every hook
I tripped up on this and thought it would be good to point this out in the docs. I'm not sure if this is the best place for it and I'm surely open to rewording. I just was here reading and cory_fu had to remind me about it while trying to figure out some code.