Supporting app level relation data (#232)
Add app level relation data functions to the library.  App level relation data is data that the leader unit of an app can publish that will be visible to all participants of the relation.  This is contrasted with the existing system which is unit level relation data where a unit can publish to another unit on the relation exclusively.  Both forms are available.

Authors: seyeong.kim@canonical.com and Cory Johns.