Fix entry points (#260)
* Use the pkg_resources for entry point iterations

The importlib modules dropped the get method. Drop the importlib
entry_points management and use the pkg_resources module with custom
entry_points function to ensure compatibility.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Change assertEquals to assertEqual

Change the assertEquals methods to assertEqual.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Drop agent parameters in tests

Drop the agent parameters in the tests. The juju version installed is
no longer 2.9.x by default.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Create juju model

Juju no longer creates a default model and a model needs to be created
as part of the integration tests.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Use charmcraft for building charms

Juju dropped support for deploying directory based charms. Build the
test charm using charmcraft.

Signed-off-by: Billy Olsen <billy.olsen@gmail.com>

* Use the pkg_resources.iter_entry_points natively

Instead of wrapping in a method for entry_points to provide a backwards
compatible dict, simply switch to using the iter_entry_points() function
directly.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Add quotes to python version

Add quotes to the python version for integration tests.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

* Specify charmcraft version in integration test

Specify the charmcraft version in the integration test.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>

---------

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>
Signed-off-by: Billy Olsen <billy.olsen@gmail.com>